Move EscapeManualToggleFeature to correct package
