+++ lib/app/features/escape_manual/domain/escape_manual_toggle.dart
@@ -1,4 +1,4 @@
-import '../../../../core/managers/modules_sevices.dart';
+import '../../../core/managers/modules_sevices.dart';
 
 class EscapeManualToggleFeature {
   EscapeManualToggleFeature({
